8344779: GenShen: Consolidate shared heap region closures into one header
